(#68) 🔄 refactor: 준비중인 서비스 컴포넌트화
